+++ src/pages/BbsCreate.jsx
@@ -8,7 +8,7 @@ import { useBbsInput } from "../variables/useBbsInput";
 
 function BbsCreate() {
   const navigate = useNavigate();
-  const {isError, error} = useSelector((state) => state.posts)
+  const { isError } = useSelector((state) => state.posts)
   const { region, category } = useSelector((state) => state.selects);
   const [inputTitle, inputTitleHandler] = useBbsInput("");
   const [inputURL, inputURLHandler] = useBbsInput("");
@@ -23,10 +23,10 @@ function BbsCreate() {
   const dispatch = useDispatch();
 
   const requestPost = (payload) => {
-      dispatch(__postPosts(payload));
-    };
-  
-  
+    dispatch(__postPosts(payload));
+  };
+
+
   console.log(isError);
 
   const newPost = {
